+++ R/splinecongruence-package.R
@@ -1,6 +1,0 @@
-#' @keywords internal
-#' @importFrom stats coef vcov logLik nobs
-"_PACKAGE"
-
-# Column names referenced via non-standard evaluation in ggplot2 aes().
-utils::globalVariables(c("x", "y", "z"))